requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
bcrypt==3.1.7
beautifulsoup4==4.9.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
fabric==2.5.0
idna==2.9
idna-ssl==1.1.0
importlib-metadata==1.6.1
invoke==1.4.1
lxml==4.5.1
more-itertools==8.4.0
multidict==4.7.5
packaging==20.4
paramiko==2.7.1
pluggy==0.13.1
py==1.8.2
pycparser==2.20
PyMySQL==0.9.3
PyNaCl==1.4.0
pyparsing==2.4.7
pytest==5.4.3
pytest-aiohttp==0.3.0
python-aiml==0.9.3
python-dotenv==0.12.0
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.17
typing-extensions==3.7.4.2
urllib3==1.25.9
wcwidth==0.2.4
yarl==1.4.2
zipp==3.1.0
